Count k-mers for string sequences (the size of k-mer should be larger than one)

Count k-mers for string sequences (the size of k-mer should be larger than one)

count_kmers_str(s, d, alphabet, pos)

Arguments

s
a string vector representing an input sequence

d
an integer vector representing gaps between consecutive elements of k-mer

alphabet
a string vector representing valid elements of k-mer

pos
a logical value that denotes whether positional k-mers should be generated

Value

a named vector with counts of k-mers

Details

K-mers that contain elements from alphabet but do not exist in the input sequence are also generated.

Examples

count_kmers_str(
c("a", "b", "c", "d", "x", "y", "z", "z", "a", "a"),
d=c(0,0),
c("a", "b", "c", "z"),
pos=FALSE)

#> c.a.z a.z.b c.a.b b.b.c z.a.c c.b.c c.z.b b.b.b c.a.a c.c.a a.a.c a.a.z b.b.a 
#>     0     0     0     0     0     0     0     0     0     0     0     0     0 
#> a.c.z c.c.b b.c.z z.c.z c.z.a c.b.z a.b.c a.z.c z.c.b a.c.c z.b.z c.a.c b.z.b 
#>     0     0     0     0     0     0     1     0     0     0     0     0     0 
#> b.a.b c.c.c z.a.a b.c.b c.z.z z.z.z z.a.z b.a.c b.c.c b.z.a a.b.z z.z.c z.c.c 
#>     0     0     1     0     0     0     0     0     0     0     0     0     0 
#> c.b.a b.b.z b.a.z z.a.b a.a.b c.z.c z.b.c a.c.b a.a.a z.b.a b.c.a c.b.b z.c.a 
#>     0     0     0     0     0     0     0     0     0     0     0     0     0 
#> a.z.a c.c.z a.b.b a.b.a z.b.b b.z.z z.z.a b.a.a a.z.z b.z.c z.z.b a.c.a 
#>     0     0     0     0     0     0     1     0     0     0     0     0