Bind rows of several k-mer matrices

The function binds rows of several input k-mer matrices (of type Matrix), which are results of count_kmers and count_multimers. This implementation also handles properly k-mer matrices that do not have the same columns, as opposed to the implementation of rbind.

rbind_columnwise(...)

Arguments

...
k-mer matrices of type Matrix

Value

a k-mer matrix of type Matrix that is the result of the rbind operation

See also

Examples

batch_size <- 1

# k-mer counting
resA <- count_kmers(c("AAAAA", "ASASSSSASSA"), k=5, batch_size=batch_size)

#> Single-threaded mode enabled. In order to speed up computations, increase defined batch_size or use a default value

resB <- count_multimers(c("HWHSHS", "AASDCASD"), k_vector=c(3, 5), batch_size=batch_size)

#> Single-threaded mode enabled. In order to speed up computations, increase defined batch_size or use a default value

# rbind
res <- rbind_columnwise(resA, resB)